==> 3.1 nMDS.Rmd <==
---
title: "3.1 nMDS"
author: "TG"
date: "2023-06-27"
output:
   prettydoc::html_pretty:
    theme: cayman
    highlight: github
---



```{r message=FALSE, warning=FALSE}

library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(stringr)
library(Hmisc)
library(vegan)
library(mobr)
library(lme4)
library(lmerTest)
library(MuMIn)
library(rareNMtests)
library(zetadiv)
library(FuzzyQ)

```



color scales definitions 

```{r}
color_op6 <- c("#20A39E","#23001E","#FFBA49","#EF5B5B")
  
save_path<-"G:\\My Drive\\PHD\\chapter 1 strom effect\\R_storm_ms\\plots\\nMDS"
model_path <-"G:\\My Drive\\PHD\\chapter 1 strom effect\\R_storm_ms\\models\\nmds"
```


# Data prepartion

```{r}
load("fish_full_data.rdata")
select <- dplyr::select
```

```{r}
fish_prep<-fish_full_data
```

#### 1. remove irregular sites


```{r}

knolls_to_remove<-c("South bridge reserve","North bridge reserve","south bridge reserve",
                    "Tamar","north bridge reserve")

fish_prep<-fish_prep %>% filter(!Knoll %in% knolls_to_remove) %>% 
filter(Site !="North beach")

```

#### 2. keep data from 2018 and on

```{r}
fish_prep<-fish_prep %>% filter(Year >2017)
```


#### 3. keep only sure identifications

filter confidence = 0


```{r}

fish_prep<-fish_prep %>% filter(Confidence <1)

```

#### 4. keep one observer from each knoll


```{r}
one_observer<-list()
loop<-1
set.seed(1)

for (i in unique(fish_prep$survey_id)) {
  
  data <- fish_prep %>% filter(survey_id == i)
  
  obs_amount <- length(unique(data$Observer))
  obs_name <- unique(data$Observer)
  chosen_obs <- sample(obs_name,1)  
 
  filter_data<-data %>% filter(Observer == chosen_obs)
  one_observer[[loop]]<-filter_data
  loop<-loop+1
  
  
}

fish_prep <- bind_rows(one_observer)

rm(data,filter_data,one_observer)



```


#### 5. convert year_month to factor

```{r}

fish_prep$year_month<-as.factor(fish_prep$year_month)

```


#### 6. keep only the knolls that were surveyed in 2020 spring (storm)

```{r}

knolls_2020<-fish_prep %>% filter(year_month %in% c("2020 a")) %>% distinct(knoll_id)

fish_prep<-fish_prep %>% filter(knoll_id %in% knolls_2020$knoll_id)

# make sure i have the right knolls number


knoll_site_year <-fish_prep %>% group_by(Site,year_month) %>% summarise("knoll_number"= n_distinct(knoll_id))

```

#### 7. use all fish 


```{r}

all_fish <- fish_prep 

```

#### 8. convert to wide data


```{r}

wide <- all_fish %>% 
  select(Site,year_month,Knoll,Species,Abundance) %>%
  group_by(Site,year_month,Knoll,Species) %>% 
    summarise(Abundance=sum(Abundance))

wide <- spread(wide,
               Species,
               Abundance,
               fill = 0)

```

# nMDS with no transformation


Seperatly for each by site

include:

* nmds plot
* Adonis
* Tukey

```{r}


tucky_res<-list()


for (i in unique(wide$Site)) {
  
  # orgenize the data
  
  data <- wide  %>% filter(Site == i)
  
  Year <- factor(data$year_month)
  
  nmds_data <- data[,4:ncol(data)] #keep only species column
  
  # create the ordination 

  ord <- metaMDS(nmds_data,distance = "bray" ,trace = F,k=2,autotransform = FALSE)   
 
  plot_name<-paste0(save_path,"\\",i,"_no_trans_all_fish",".pdf")
   
 # save plot  ---start  

pdf(file = plot_name,   # The directory you want to save the file in
    width = 5,          # The width of the plot in inches
    height = 5) 

ordiplot(ord, type = "n",main = paste(i,", stress=",round(ord$stress,3)))
orditorp(ord, label = T, display = "sites", col = color_op6[Year],pch = 16)
ordiellipse(ord, groups = data$year_month,kind = "sd", draw = "polygon",alpha = 0.35,label=F ,lty = 1,col = color_op6)
legend("bottomright", legend = levels(Year), bty = "n", col = color_op6, pch = 15,cex=1.5)

dev.off() #   --- end

   
   print(i)
   
   # Adonis 

   dis <- vegdist(nmds_data,method="bray")
   adon.results<-adonis(dis ~  year_month,data = data, method="bray")
   
   print(adon.results$aov.tab) 
   print("adonis end")


   # tukeyHSD

    mod <- betadisper(dis,data$year_month)

    res<-TukeyHSD(mod)
    res<- as.data.frame(res[["group"]])
    res$sample <- rownames(res)


tucky_res[[i]]<-res

}

openxlsx::write.xlsx(tucky_res, file = paste0(model_path,'\\tucky_res.xlsx'))
```

* adonis - test differences in centroids and data dispersion

https://chrischizinski.github.io/rstats/adonis/
