==> 4.2 zeta figure for each site.Rmd <==
---
title: "4.2 zeta figure for each site"
author: "TG"
date: "2023-06-27"
output:
   prettydoc::html_pretty:
    theme: cayman
    highlight: github
---


*In this script ill concentrate all the models and plot constructing the Zeta diversity figure*

libraries

```{r message=FALSE, warning=FALSE}

library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(stringr)
library(Hmisc)
library(vegan)
library(mobr)
library(lme4)
library(lmerTest)
library(MuMIn)
library(Rarefy)
library(rareNMtests)
library(zetadiv)

```

color scales definitions 

```{r}
color_op6 <- c("#20A39E","#23001E","#FFBA49","#EF5B5B")

save_path<-"G:\\My Drive\\PHD\\chapter 1 strom effect\\R_storm_ms\\plots\\Zeta\\each site"
```


# Data prepartion

```{r}
load("fish_full_data.rdata")
select <- dplyr::select
```

## general steps:

```{r}

fish_prep<-fish_full_data

```

2. keep only 2018 - 2021

```{r}

fish_prep<- fish_prep %>% filter(Year>2017)

```


3. remove irregular sites


```{r}

knolls_to_remove<-c("South bridge reserve","North bridge reserve","south bridge reserve",
                    "Tamar","north bridge reserve")

fish_prep<-fish_prep %>% filter(!Knoll %in% knolls_to_remove) %>% 
filter(Site !="North beach")

```


3*. can it be that patterns derive from different knolls number in the diffrent year?
im adding here additional filter layer that keeps only the knolls that were survey in 2020 storm

```{r}
knolls_strom<-fish_prep %>% filter(year_month %in% c("2020 a")) %>% distinct(knoll_id)

fish_prep<-fish_prep %>% filter(knoll_id %in% knolls_strom$knoll_id)

```

4. keep only sure identifications

filter confidence = 0

```{r}

fish_prep<-fish_prep %>% filter(Confidence <1)

```

5. keep one observer from each knoll

```{r}
one_observer<-list()
loop<-1
set.seed(1)

for (i in unique(fish_prep$survey_id)) {
  
  data <- fish_prep %>% filter(survey_id == i)
  
  obs_amount <- length(unique(data$Observer))
  obs_name <- unique(data$Observer)
  chosen_obs <- sample(obs_name,1)  
 
  filter_data<-data %>% filter(Observer == chosen_obs)
  one_observer[[loop]]<-filter_data
  loop<-loop+1
  
  
}

fish_prep <- bind_rows(one_observer)

rm(data,filter_data,one_observer)

# make sure i have the right knolls number


knoll_site_year <-fish_prep %>% group_by(Site,year_month) %>% summarise("knoll_number"= n_distinct(knoll_id))

```


6. convert year_month to factor

```{r}

fish_prep$year_month<-as.factor(fish_prep$year_month)

```

7. remove knoll without coordinates


```{r}
fish_prep <- fish_prep %>% drop_na(DD_X)
```

8. add id column for combination of site + year + month 

```{r}

fish_prep$site_sample<-paste(fish_prep$Site,
                             fish_prep$Year,
                             month(fish_prep$Date),
                             sep=" ")
```

9. keep resident fish only


```{r}

resident_fish <- fish_prep %>% filter(Status != "T")

```


10. convert to wide format


resident fish

```{r}
resident_fish <- resident_fish %>% 
  select(site_sample,Site,year_month,Knoll,DD_X,DD_Y,Species,Abundance) %>%
  group_by(site_sample,Site,year_month,Knoll,DD_X,DD_Y,Species) %>% 
    summarise(Abundance=sum(Abundance))

resident_fish <- spread(resident_fish,
                      Species,
                      Abundance,
                      fill = 0)

```

# zeta calculation

## Resident fish


```{r}

zeta_data_resident_list<-list()
zeta_resident_list<-list()
aic_list_resident<-list()
loop_resident<-1


resident_fish<- resident_fish%>% mutate(loop_id = paste(Site,year_month,sep = "_"),.after = 1)

for (i in unique(resident_fish$loop_id)) {

# filter one year (year_month) 
      
  one_data <- resident_fish %>% filter(loop_id == i )

# keep recored of the site and year_month 
  
  t_year_month<-unique(one_data$year_month)
  t_Site<-unique(one_data$Site)
  

# data only of the species
  
one_species<-one_data[,8:ncol(one_data)]
one_species[one_species>0]<-1 # Convert to presence/absence
one_species<-as.data.frame(one_species)

one_species<-one_species%>% select(where(~ is.numeric(.) && sum(.) >= 1)) #remove 0 columns

# data for the coords only 

one_coord <- one_data[,6:7]
one_coord <- one_coord %>% rename(x = DD_X,y = DD_Y)
one_coord <- as.data.frame(one_coord)

# Compute the zeta diversity

zeta <- Zeta.decline.mc(one_species,
                one_coord,
                orders = 1:nrow(one_species),
                sam = 1000,
                NON = T,
                plot = F)

# save the results into data frame

data <- data.frame("z_order"=zeta[["zeta.order"]],
                   "zeta_value"=zeta[["zeta.val"]],
                   "zeta_sd"=zeta[["zeta.val.sd"]],
                   "zeta_ratio"=c(NA,zeta[["ratio"]]),
                   "year_month"=rep(t_year_month),
                   "Site"= rep(t_Site))

# create the aic data frame 


 aic_table<-as.data.frame(zeta[["aic"]])
 aic_table$year_month<- rep(t_year_month)
 aic_table$method<-rownames(aic_table)
 
 aic_table$method[aic_table$method=="zeta$zeta.exp"]<-"exp"
 aic_table$method[aic_table$method=="zeta$zeta.pl"]<-"plog" 
 
# calculte the delta and keep only the best model
 
 aic_table <- aic_table %>% mutate("delta_aic"= min(AIC)-max(AIC)) %>%  filter(AIC == min(AIC))
 
 # save the results 
   
 aic_list_resident[[loop_resident]]<-aic_table
 zeta_resident_list[[i]]<-zeta
 zeta_data_resident_list[[loop_resident]]<-data


 # move to next year
  
loop_resident<-loop_resident+1

}
  
zeta_divesity_resident<-bind_rows(zeta_data_resident_list)

aic_table_resident<-bind_rows(aic_list_resident)


```

### plot

#### zeta decline


```{r}
for (i in unique(zeta_divesity_resident$Site)) {
  
  site_zeta <- zeta_divesity_resident %>% filter(Site == i)
  
  ggplot(data =site_zeta,
       aes(x=z_order,y=zeta_value,color =year_month))+
  geom_line(size=0.5)+
  geom_point(size = 4)+
  theme_classic()+ ggtitle(i)+
  xlab("Zeta order")+
  ylab("Zeta diversity")+
  theme(text = element_text(family = "serif"),
        axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =16),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        plot.title = element_text(size=24))+
  scale_x_continuous(limits = c(0,16), breaks = seq(0,16, by = 4))+
  scale_color_manual(values = color_op6,
                     name = "Year",
                     labels = c("PreS1","PostS1","PostS2" ,
                                "PostS3"))
save_plot <-paste(i,"_","zeta_resident_fish_2018.png")
save_plot_pdf <- paste(i,"_","zeta_resident_fish_2018.pdf")

ggsave(save_plot, width = 10, height = 5, dpi = 300,path = save_path) 

ggsave(save_plot_pdf, width = 10, height = 5, dpi = 300,path = save_path)


}

```

#### zeta ratio

```{r}

for (i in unique(zeta_divesity_resident$Site)) {
  
  site_ratio <- zeta_divesity_resident %>% filter(Site == i)
  
ggplot(data =site_ratio,
       aes(x=z_order,y=zeta_ratio,color =year_month))+
  geom_line(size=0.5)+
  geom_point(size=4)+
  theme_classic()+ ggtitle(i)+
  xlab("Zeta order")+
  ylab("Zeta ratio")+
  theme(text = element_text(family = "serif"),
        axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =16),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        plot.title = element_text(size=24))+
    scale_color_manual(values = color_op6,name = "Year",
                           labels = c(                                  "PreS1","PostS1", "PostS2" ,
                                      "PostS3"))


save_plot <-paste(i,"_","zeta_resident_fish_ratio_2018.png")
save_plot_pdf <- paste(i,"_","zeta_resident_fish_ratio_2018.pdf")

ggsave(save_plot,
       width = 10, height = 5, dpi = 300,
       path =save_path )

ggsave(save_plot_pdf,
       width = 10, height = 5, dpi = 300,
       path =save_path)

}








```


# create null model

1. How many species in 2020 a in each knoll


```{r}

# how many species was present im 2020a

unique_sp_storm <- fish_prep %>%
  filter(year_month == "2020 a",Status == "C") %>% 
  group_by(knoll_id) %>% summarise("sp_to_sample"=n_distinct(Species))


null_prep <- fish_prep %>% filter(year_month == "2018 b",Status == "C") %>%  
  select(Site,year_month,knoll_id,DD_X,DD_Y,Species,Abundance) %>%
  group_by(Site,year_month,knoll_id,DD_X,DD_Y,Species) %>% 
  summarise(Abundance=sum(Abundance))

null_prep<-null_prep %>% group_by(Site,year_month,knoll_id,DD_X,DD_Y) %>% 
  mutate(sp_in_knoll=n_distinct(Species)) %>% ungroup()

null_prep<-left_join(null_prep,unique_sp_storm)


null_prep$sp_to_sample[is.na(null_prep$sp_to_sample)]<-0

# the number of knoll in which there is more species in 2020

nrow(null_prep %>% distinct(knoll_id,.keep_all = T) %>% filter(sp_to_sample>sp_in_knoll))


null_prep$sp_to_sample<-ifelse(null_prep$sp_to_sample>null_prep$sp_in_knoll,
                               null_prep$sp_in_knoll,
                               null_prep$sp_to_sample)

```



2. draw randomly the same number of species as 2020a from 2018 for each knoll and itrate it 100 times


```{r}


itr<-list()


pb <- txtProgressBar(min = 0,max = 100,style = 3)

for (j in 1:100) {

  null <- list()

 for (i in unique(null_prep$knoll_id)) {
  
  knoll<-null_prep %>% filter(knoll_id == i)
  
  knoll$Abundance<-rep(0)
  
  n_draw<-unique(knoll$sp_to_sample) # how many lines to choose
  
  n_lines <- nrow(knoll) # how many lines
  
  lines_to_fill <- sample(n_lines,n_draw) # sample lines according to n_draw
  
  knoll$Abundance[lines_to_fill]<-1 # in the chosen lines change abundance to be 1
  
  
  knoll<- knoll  %>% 
    select(-sp_in_knoll,-year_month,-sp_to_sample)
  
  null[[i]]<-knoll


}

null<-bind_rows(null) 
  
null <- null %>% mutate(itr = rep(j),.before = 1)

null<- spread(null,Species,Abundance,fill = 0)

itr[[j]]<- null

setTxtProgressBar(pb, j)
  
}

close(pb)

itr<-bind_rows(itr)


```

3. create zeta for each itr


```{r}

zeta_null<-list()

itr<-itr %>% mutate(itr_site = paste(itr,Site,sep = "_"),.after = 1)



for (i in unique(itr$itr_site)) {

# filter one year (year_month) 
      
  one_data <- itr %>% filter(itr_site == i )

# keep recored of the site and year_month 
  
  t_itr<-unique(one_data$itr)
  t_Site<-unique(one_data$Site)
  

# data only of the species
  
one_species<-one_data[,7:ncol(one_data)]
one_species<-as.data.frame(one_species)

# data for the coords only 

one_coord <- one_data[,5:6]
one_coord <- one_coord %>% rename(x = DD_X,y = DD_Y)
one_coord <- as.data.frame(one_coord)

# Compute the zeta diversity

zeta <- Zeta.decline.mc(one_species,
                one_coord,
                orders = 1:nrow(one_species),
                sam = 100,
                NON = T,
                plot = T)

# save the results into data frame

data <- data.frame("z_order"=zeta[["zeta.order"]],
                   "zeta_value"=zeta[["zeta.val"]],
                   "zeta_sd"=zeta[["zeta.val.sd"]],
                   "zeta_ratio"=c(NA,zeta[["ratio"]]),
                   "itr"=rep(t_itr),
                   "Site" = rep(t_Site))

 # save the results 
   

 zeta_null[[i]]<-data

 setTxtProgressBar(pb, i)

}

  
zeta_null<-bind_rows(zeta_null)


```


4. group by zeta order


```{r}
null_zeta_decay<- zeta_null %>%
  group_by(z_order,Site) %>%
  summarise(mean_zeta_value = mean(zeta_value),
            zeta_sd = sd(zeta_value))


```

5. group by zeta ratio


```{r}

null_zeta_ratio<- zeta_null 
  
null_zeta_ratio$zeta_ratio[null_zeta_ratio$zeta_ratio == "NaN"]<-0 

null_zeta_ratio<-null_zeta_ratio %>% 
group_by(z_order,Site) %>%
  summarise(mean_zeta_ratio = mean(zeta_ratio,na.rm = F))



```

output from species null model:

1.null_zeta_decay - Zeta decline
2. null_zeta_ratio - Zeta Ratios


## null model by individuals


```{r}
# how many species was present im 2020a

ind_storm <- fish_prep %>%
  filter(year_month == "2020 a",Status == "C") %>% 
  group_by(knoll_id) %>% summarise("sp_to_sample"=n_distinct(Species),"ind_to_sample"=sum(Abundance))

# uncount so each line is one individual

ind_null_prep <- fish_prep %>% filter(year_month == "2018 b",Status == "C") %>%  
  select(Site,year_month,knoll_id,DD_X,DD_Y,Species,Abundance) %>%
  group_by(Site,year_month,knoll_id,DD_X,DD_Y,Species) %>% uncount(Abundance) %>% mutate(Abundance = rep(0)) 

 
# note - abundance is now a dummy variable that will be useful only in the nest stages....


ind_null_prep<-ind_null_prep %>% group_by(Site,year_month,knoll_id,DD_X,DD_Y) %>% 
  mutate(ind_in_knoll=n()) %>% ungroup()

ind_null_prep<-left_join(ind_null_prep,ind_storm)

# in knoll caves_5 there was no cryptic fish at 2020 a

ind_null_prep$sp_to_sample[is.na(ind_null_prep$sp_to_sample)]<-0
ind_null_prep$ind_to_sample[is.na(ind_null_prep$ind_to_sample)]<-0

# the number of knoll in which there is more ind in 2020

nrow(ind_null_prep %>% distinct(knoll_id,.keep_all = T) %>% filter(ind_to_sample>ind_in_knoll))


ind_null_prep$ind_to_sample<-ifelse(ind_null_prep$ind_to_sample>ind_null_prep$ind_in_knoll,
                               ind_null_prep$ind_in_knoll,
                               ind_null_prep$ind_to_sample)

```

2. draw randomly the same number of individuals as 2020a from 2018 for each knoll and itrate it 100 times


```{r}

ind_itr<-list()


pb <- txtProgressBar(min = 0,max = 100,style = 3)

for (j in 1:100) {

  ind_null <- list()

 for (i in unique(ind_null_prep$knoll_id)) {
  
  knoll<-ind_null_prep %>% filter(knoll_id == i)
  
  n_draw<-unique(knoll$ind_to_sample) # how many lines to choose
  
  n_lines <- nrow(knoll) # how many lines
  
  lines_to_fill <- sample(n_lines,n_draw) # sample lines according to n_draw
  
  knoll$Abundance[lines_to_fill]<-1 # in the chosen lines change abundance to be 1
  
  knoll<- knoll  %>% 
    select(-ind_in_knoll,-year_month,-sp_to_sample,-ind_to_sample)
  
  knoll<-knoll %>% group_by_at(vars(-Abundance)) %>% summarise(Abundance=sum(Abundance))
  
  ind_null[[i]]<-knoll


}

ind_null<-bind_rows(ind_null) 
  
ind_null <- ind_null %>% mutate(itr = rep(j),.before = 1)

ind_null<- spread(ind_null,Species,Abundance,fill = 0)

ind_itr[[j]]<- ind_null

setTxtProgressBar(pb, j)
  
}

close(pb)

ind_itr<-bind_rows(ind_itr)
```

3. create zeta for each itr


```{r}


ind_itr<- ind_itr %>% mutate(loop_id = paste(itr,Site,sep = "_"),.after = 2)

ind_zeta_null<-list()

for (i in unique(ind_itr$loop_id)) {

# filter one year (year_month) 
      
  one_data <- ind_itr %>% filter(loop_id == i )

# keep recored of the itr
  
  t_itr<-unique(one_data$itr)
  t_Site<-unique(one_data$Site)
  

# data only of the species
  
one_species<-one_data[,7:ncol(one_data)]
one_species[one_species>0]<-1 # Convert to presence/absence
one_species<-as.data.frame(one_species)
one_species<-one_species%>% select(where(~ is.numeric(.) && sum(.) >= 1)) #remove 0 columns

one_species<-as.data.frame(one_species)


# data for the coords only 

one_coord <- one_data[,5:6]
one_coord <- one_coord %>% rename(x = DD_X,y = DD_Y)
one_coord <- as.data.frame(one_coord)

# Compute the zeta diversity

zeta <- Zeta.decline.mc(one_species,
                one_coord,
                orders = 1:nrow(one_species),
                sam = 100,
                NON = T,
                plot = T)

# save the results into data frame

data <- data.frame("z_order"=zeta[["zeta.order"]],
                   "zeta_value"=zeta[["zeta.val"]],
                   "zeta_sd"=zeta[["zeta.val.sd"]],
                   "zeta_ratio"=c(NA,zeta[["ratio"]]),
                   "site" = rep(t_Site),
                   "itr"=rep(t_itr))

 # save the results 
   

 ind_zeta_null[[i]]<-data



}


  
ind_zeta_null<-bind_rows(ind_zeta_null)

```

4. group by zeta order


```{r}
ind_null_zeta_decay<- ind_zeta_null %>%
  group_by(z_order,site) %>%
  summarise(mean_zeta_value = mean(zeta_value),
            zeta_sd = sd(zeta_value))

```

```{r}

zeta_divesity_resident_short<-zeta_divesity_resident %>%
filter(year_month %in% c("2020 a","2018 b"))

sites<-unique(zeta_divesity_resident_short$Site)

for (i in unique(sites)) {
  
  real_data <-zeta_divesity_resident_short  %>%
    filter(Site ==i)
  
  null_1<- ind_null_zeta_decay %>% filter(site == i)
  
  null_2<- null_zeta_decay %>% filter(Site == i)
  
  
  ggplot(data =real_data,
       aes(x=z_order,y=zeta_value,color =year_month))+
  geom_line(size=0.5)+
  geom_point(size=4)+
  theme_classic()+ ggtitle(i)+
  xlab("Zeta order")+
  ylab("Zeta diversity")+
  theme(text = element_text(family = "serif"),
        axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =16),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        plot.title = element_text(size=24))+
  scale_x_continuous(limits = c(0,16), breaks = seq(0,16, by = 4))+
  scale_color_manual(values = color_op6[1:2],
                     name = "Year",
                     labels = c("PreS1","PostS1"))+
  geom_line(data = null_1,
            aes(x = z_order,y= mean_zeta_value),
            color = "#bc201c",
            linetype = "dashed",
            size=0.5) + 
  geom_point(data = null_1,
             aes(x = z_order,y= mean_zeta_value),
             color = "#bc201c",
             size=3,
             shape = 17)+
  geom_line(data = null_2,
            aes(x = z_order,y= mean_zeta_value),
            color = "#bc201c",
            linetype = "dashed",
            size=0.5) +
  geom_point(data = null_2,
             aes(x = z_order,y= mean_zeta_value),
             color = "#bc201c",
             size=3,
             shape = 15)
  
save_plot <-paste(i,"_","null_zeta_decay.png")
save_plot_pdf <- paste(i,"_","null_zeta_decay.pdf")
  
ggsave(save_plot, width = 10, height = 5, dpi = 300,path = save_path)

ggsave(save_plot_pdf, width = 10, height = 5, dpi = 300,path = save_path)  
  
}

```


```{r}
ind_null_zeta_ratio<- ind_zeta_null 
  
ind_null_zeta_ratio$zeta_ratio[ind_null_zeta_ratio$zeta_ratio == "NaN"]<-0  

ind_null_zeta_ratio<-ind_null_zeta_ratio %>% 
group_by(z_order,site) %>%
  summarise(mean_zeta_ratio = mean(zeta_ratio))


```


```{r}

for (i in unique(sites)) {
  
 data <-zeta_divesity_resident_short %>% filter(Site == i ) 

 null_1<-null_zeta_ratio %>% filter(Site == i)
 
 null_2 <- ind_null_zeta_ratio %>% filter(site == i)
 
 
ggplot(data =data,
       aes(x=z_order,y=zeta_ratio,color =year_month))+
  geom_line(size=0.5)+
  geom_point(size=4)+
  theme_classic()+ ggtitle(i)+
  xlab("Zeta order")+
  ylab("Zeta ratio")+
  theme(text = element_text(family = "serif"),
        axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =16),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        plot.title = element_text(size=24))+
  scale_color_manual(values = color_op6[1:2],name = "Year",
                     labels = c("PreS1","PostS1"))+
  geom_line(data = null_1,
            aes(x = z_order,y= mean_zeta_ratio),
            linetype = "dashed",
            color = "#bc201c",
            size=0.5)+
  geom_point(data = null_1,
             aes(x = z_order,y= mean_zeta_ratio),
             color = "#bc201c",
             size=3,
             shape=15)+
  geom_line(data = null_2,
            aes(x = z_order,y= mean_zeta_ratio),
            linetype = "dashed",
            color = "#bc201c",
            size=0.5)+
  geom_point(data = null_2,
             aes(x = z_order,y= mean_zeta_ratio),
             color = "#bc201c",
             size=3,
             shape = 17)

save_plot <-paste(i,"_","null_ratio.png")
save_plot_pdf <- paste(i,"_","null_ratio.pdf")

ggsave(save_plot,
       width = 10, height = 5, dpi = 300,
       path = save_path)

ggsave(save_plot_pdf,
       width = 10, height = 5, dpi = 300,
       path = save_path) 

}

```
